FXC-1718 Fix sorting for `//*[___]` type of queries.
This was caused by the //* being rewritten to
/descendant-or-self::node()/child::element(). The
descendant-or-self::node() part outputted all the nodes, the child::
part transformed it, simulating a breadth-first search, which did not
order parents nicely to the children.

By sorting them, we can use a O(N*log(N)) search. We really should
rewrite //XXX to /descendant::XX though, because that has a O(N)
complexity.