Implement multi-core fuzzer (#348)
Implement multi-core fuzzer

Refactor the fuzzer implementation into the engine and the
corpus_based_fuzzer components. This allowes multiple fuzzers
to co-exist side-by-side.

The engine is overhauled to support running on multiple core in
parallel. Big shoutout to @mvanbutselaar for his help with the
implementation. The output of the engine is colorized for improved
readability.

Co-authored-by: codefactor-io <support@codefactor.io>
Co-authored-by: Martin Middel <martin.middel@fontoxml.com>