Keep reference equality roundtripping maps and arrays through XPath
This can improve performance because a map does not have to be copied a second time when returning
from XPath. It also makes the APIs easier to use because an entry in the map will be exactly the
same as the entry that was put in to engine as a variable.

This makes FontoXPath more predictable