Simple implementation of replace() excluding flags argument
Like the implementation of tokenize(), use JavaScript RegExp directly. This
means that the implementation will be wrong in all the ways that
XPath/XQuery regexs differ from JavaScript. In particular, the XPath syntax
for backreferences won’t work, and the preferred match for alternatives will
be wrong in many cases.

On the other hand, all the tests from
https://www.w3.org/TR/xpath-functions-31/#func-replace pass, except those
involving backreferences.

The flags argument (fourth argument) is not supported.