Fix annotation in the built version of fontoxpath being skipped
The editor performance was not improved when we introduced the new annotation feature in
fontoxpath. Turns out this was happening because the closude compiler was rewriting the Object maps
that are being used to register the annotation functions.

By replacing them with ES6 Maps, this just works. This should improve the performance of the editor.