Fix annotation in the built version of fontoxpath being skipped
* Fix annotation in the built version of fontoxpath being skipped

The editor performance was not improved when we introduced the new annotation feature in
fontoxpath. Turns out this was happening because the closude compiler was rewriting the Object maps
that are being used to register the annotation functions.

By replacing them with ES6 Maps, this just works. This should improve the performance of the editor.

* Remove unused dependency on prettier from demo

* Fix attribute tests with URILiteral expressions

The Q{} syntax can also be used for attributes, in which case prefixes do not matter.