Fix bug where attributes are compared with literals
Attributes get annotated as attributes and each comparison between an
attribute and an interger literal automatically gets accepted my the
annotator. This means that the shortcut without additional type checks
gets executed so there is no error that gets thrown.

This was solved by adding a check in Compare.ts to make sure the actual
types are the same as the types the annotated "guessed".