Fix text() not selecting CDATA sections
While the expression version of the text() node test did check for the
CDataSection node type, the codegen-emitted code did not. Additionally,
the bucket generated for a CDataSection node was type-4, while the
text() selector always uses type-3. This fixes both issues and adds a
few tests to confirm that fix.